chore: move functions to proper location
