+++ apps/client/src/plugins/LayerSwitcher/LayerSwitcherProvider.js
@@ -12,6 +12,7 @@ import LayerSwitcherView from "./LayerSwitcherView.js";
 import { useLayerZoomWarningSnackbar } from "./useLayerZoomWarningSnackbar";
 import { functionalOk as functionalCookieOk } from "../../models/Cookie";
 import LocalStorageHelper from "../../utils/LocalStorageHelper";
+import { setOLSubLayers, getAllLayerIdsInGroup } from "../../utils/groupLayers";
 
 export const QUICK_ACCESS_KEY = "quickAccess";
 export const QUICK_ACCESS_LS_KEY = "quickAccessLayers";
@@ -94,54 +95,6 @@ const LayerZoomVisibleSnackbarProvider = ({ children, layers }) => {
       {children}
     </>
   );
-};
-
-export const setOLSubLayers = (olLayer, visibleSubLayersArray) => {
-  if (visibleSubLayersArray.length === 0) {
-    // Fix underlying source
-    olLayer.getSource().updateParams({
-      // Ensure that the list of sublayers is emptied (otherwise they'd be
-      // "remembered" the next time user toggles group)
-      LAYERS: "",
-      // Remove any filters
-      CQL_FILTER: null,
-    });
-
-    // Hide the layer in OL
-    olLayer.setVisible(false);
-  } else {
-    // Set LAYERS and STYLES so that the exact sublayers that are needed
-    // will be visible
-    olLayer.getSource().updateParams({
-      // join(), so we always provide a string as value to LAYERS
-      LAYERS: visibleSubLayersArray.join(),
-      // Filter STYLES to only contain styles for currently visible layers,
-      // and maintain the order from layersInfo (it's crucial that the order
-      // of STYLES corresponds exactly to the order of LAYERS!)
-      STYLES: Object.entries(olLayer.layersInfo)
-        .filter((k) => visibleSubLayersArray.indexOf(k[0]) !== -1)
-        .map((l) => l[1].style)
-        .join(","),
-      CQL_FILTER: null,
-    });
-    olLayer.set("subLayers", visibleSubLayersArray);
-    olLayer.setVisible(true);
-  }
-};
-
-// TODO move to common. Copied from LayerGroup.js
-const getAllLayerIdsInGroup = (group) => {
-  if (!group) {
-    return [];
-  }
-
-  if (!group.children) {
-    return [group.id];
-  } else {
-    return group.children.flatMap((c) => {
-      return getAllLayerIdsInGroup(c);
-    });
-  }
 };
 
 const getGroupConfigById = (tree, groupId) => {
